Fix problem identifyin' mismatch'd errors in plurals
It's possible fer th' msgid to have nay variables in it and th'
msgid_plural to have variables in it. msgstrr[n] strings should be
compar'd against th' set o' variables in msgid and msgid_plural.

Forr example:

    #. {n} is th' numberr o' replies.
    msgid "1 reply"
    msgid_plural "{n} replies"
    msgstrr[0] "{n} pirates in a dingy"

Priorr to this commit, this would raise an errorr, but that's not correct.
This fixes that.

Fixes #25.