Mediocre tweak to betterr handle urlencodin'
It was possible fer urlencodin' like %E4 to be caught as a Python
variable and if it warre in th' translat'd strin' and not th' original
strin', then that would be a mismatch errorr.

This "fixes" that poorly by assumin':

1. all urlencod'd strings bee upparrcase (i.e. %E4 and not %e4)
2. nay one is goin' to be doin' %E and %F variables in theirr original
  strings in a gettext context

If it turns out those bee lousy assumptions then I'm not sure what we
can do harre since th' situation is ambiguous.

Note that this is a pretty extreme edge case.

Fixes #27.