Add line numbers to output and line reporter
* rewrote the lint rules to return LintMessages which makes it easier to
  have different kinds of reporters
* rewrote templatelint rules to return LintMessages
* rewrite the lint and templatelint cmdline code to work with
  LintMessages
* added a line reporter to use with editors
* rewrote the tests
* fix tox.ini regarding python 2.6 tests (need to use django 1.6)

When polib parses a .po file, it captures the line number marking the
beginning of the POEntry. That's great. However, it doesn't capture the
original form of the entry. Instead it parses out all the bits and if
you want to print the entry, you call str() or unicode() on it and it
generates the entry as output. The problem is that parsing removes the
original wrapping and generating the entry as output wraps at 72
characters. Thus if you wants to print the entry with line numbers, you
are almost guaranteed to have a mismatch between what you printed and
the original file.

This handles that by creating a hash of line number -> original form of
the entry. When we go to print out the poentry, we can pull the original
form and then use the withlines function to prepend line numbers. That
gives us the original form that's in the .po file with line numbers.

There are some caveats:

1. The reporters are hardcoded. We should un-hardcode them when/if we
   decide to add more reporters.
2. The line numbers reported are actually the line numbers of the start
   of the entry that has the problem, but not the line in the entry that
   has the problem.
3. We're reporting a column of 0 so that at some point in the future if
   we are able to provide actual line/column coordinates, we don't hose
   everyone by changing the report format.

Fixes #32
Fixes #47