Reformat code with black and lint with flake8
This adds black to the requirements, adds a lint environment to tox to
check formatting, and runs black on the codebase to reformat it.

This also adds flake8 to the requirements, adds it to the lint tox
environment, adds a Makefile rule for it, and runs it across the
codebase.

This also fixes .editorconfig max line length settings.